my commit

+++ database/migrations/2020_09_22_220420_create_products_table.php
@@ -17,12 +17,12 @@ class CreateProductsTable extends Migration
             Schema::create('products', function (Blueprint $table) {
                 $table->bigIncrements('id');
                 $table->string('name');
-                $table->integer('purchasing_price')->nullable();
-                $table->integer('selling_price')->nullable();
-                $table->integer('total_price')->nullable();
+                $table->double('purchasing_price')->nullable();
+                $table->double('selling_price')->nullable();
+                $table->double('total_price')->nullable();
                 $table->integer('quantity')->nullable();
                 $table->string('code')->nullable();
-                $table->integer('discount')->nullable();
+                $table->double('discount')->nullable();
                 $table->unsignedBigInteger('user_id')->nullable();
                 $table->unsignedBigInteger('category_id')->nullable();
                 $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade')->onUpdate("cascade");
